Password Checker package created
